+++ mvrx/src/main/kotlin/com/airbnb/mvrx/BaseMvRxViewModel.kt
@@ -328,6 +328,129 @@ abstract class BaseMvRxViewModel<S : MvRxState>(
         .distinctUntilChanged()
         .subscribeLifecycle(owner, uniqueOnly) { (a, b, c, d) -> subscriber(a, b, c, d) }
 
+    /**
+     * Subscribe to state changes for five properties.
+     */
+    protected fun <A, B, C, D, E> selectSubscribe(
+            prop1: KProperty1<S, A>,
+            prop2: KProperty1<S, B>,
+            prop3: KProperty1<S, C>,
+            prop4: KProperty1<S, D>,
+            prop5: KProperty1<S, E>,
+            subscriber: (A, B, C, D, E) -> Unit
+    ) = selectSubscribeInternal(null, prop1, prop2, prop3, prop4, prop5, false, subscriber)
+
+    @RestrictTo(RestrictTo.Scope.LIBRARY)
+    fun <A, B, C, D, E> selectSubscribe(
+            owner: LifecycleOwner,
+            prop1: KProperty1<S, A>,
+            prop2: KProperty1<S, B>,
+            prop3: KProperty1<S, C>,
+            prop4: KProperty1<S, D>,
+            prop5: KProperty1<S, E>,
+            uniqueOnly: Boolean = false,
+            subscriber: (A, B, C, D, E) -> Unit
+    ) = selectSubscribeInternal(owner, prop1, prop2, prop3, prop4, prop5, uniqueOnly, subscriber)
+
+    private fun <A, B, C, D, E> selectSubscribeInternal(
+            owner: LifecycleOwner?,
+            prop1: KProperty1<S, A>,
+            prop2: KProperty1<S, B>,
+            prop3: KProperty1<S, C>,
+            prop4: KProperty1<S, D>,
+            prop5: KProperty1<S, E>,
+            uniqueOnly: Boolean,
+            subscriber: (A, B, C, D, E) -> Unit
+    ) = stateStore.observable
+            .map { MvRxTuple5(prop1.get(it), prop2.get(it), prop3.get(it), prop4.get(it), prop5.get(it)) }
+            .distinctUntilChanged()
+            .subscribeLifecycle(owner, uniqueOnly) { (a, b, c, d, e) -> subscriber(a, b, c, d, e) }
+
+    /**
+     * Subscribe to state changes for six properties.
+     */
+    protected fun <A, B, C, D, E, F> selectSubscribe(
+            prop1: KProperty1<S, A>,
+            prop2: KProperty1<S, B>,
+            prop3: KProperty1<S, C>,
+            prop4: KProperty1<S, D>,
+            prop5: KProperty1<S, E>,
+            prop6: KProperty1<S, F>,
+            subscriber: (A, B, C, D, E, F) -> Unit
+    ) = selectSubscribeInternal(null, prop1, prop2, prop3, prop4, prop5, prop6, false, subscriber)
+
+    @RestrictTo(RestrictTo.Scope.LIBRARY)
+    fun <A, B, C, D, E, F> selectSubscribe(
+            owner: LifecycleOwner,
+            prop1: KProperty1<S, A>,
+            prop2: KProperty1<S, B>,
+            prop3: KProperty1<S, C>,
+            prop4: KProperty1<S, D>,
+            prop5: KProperty1<S, E>,
+            prop6: KProperty1<S, F>,
+            uniqueOnly: Boolean = false,
+            subscriber: (A, B, C, D, E, F) -> Unit
+    ) = selectSubscribeInternal(owner, prop1, prop2, prop3, prop4, prop5, prop6, uniqueOnly, subscriber)
+
+    private fun <A, B, C, D, E, F> selectSubscribeInternal(
+            owner: LifecycleOwner?,
+            prop1: KProperty1<S, A>,
+            prop2: KProperty1<S, B>,
+            prop3: KProperty1<S, C>,
+            prop4: KProperty1<S, D>,
+            prop5: KProperty1<S, E>,
+            prop6: KProperty1<S, F>,
+            uniqueOnly: Boolean,
+            subscriber: (A, B, C, D, E, F) -> Unit
+    ) = stateStore.observable
+            .map { MvRxTuple6(prop1.get(it), prop2.get(it), prop3.get(it), prop4.get(it), prop5.get(it), prop6.get(it)) }
+            .distinctUntilChanged()
+            .subscribeLifecycle(owner, uniqueOnly) { (a, b, c, d, e, f) -> subscriber(a, b, c, d, e, f) }
+
+    /**
+     * Subscribe to state changes for seven properties.
+     */
+    protected fun <A, B, C, D, E, F, G> selectSubscribe(
+            prop1: KProperty1<S, A>,
+            prop2: KProperty1<S, B>,
+            prop3: KProperty1<S, C>,
+            prop4: KProperty1<S, D>,
+            prop5: KProperty1<S, E>,
+            prop6: KProperty1<S, F>,
+            prop7: KProperty1<S, G>,
+            subscriber: (A, B, C, D, E, F, G) -> Unit
+    ) = selectSubscribeInternal(null, prop1, prop2, prop3, prop4, prop5, prop6, prop7, false, subscriber)
+
+    @RestrictTo(RestrictTo.Scope.LIBRARY)
+    fun <A, B, C, D, E, F, G> selectSubscribe(
+            owner: LifecycleOwner,
+            prop1: KProperty1<S, A>,
+            prop2: KProperty1<S, B>,
+            prop3: KProperty1<S, C>,
+            prop4: KProperty1<S, D>,
+            prop5: KProperty1<S, E>,
+            prop6: KProperty1<S, F>,
+            prop7: KProperty1<S, G>,
+            uniqueOnly: Boolean = false,
+            subscriber: (A, B, C, D, E, F, G) -> Unit
+    ) = selectSubscribeInternal(owner, prop1, prop2, prop3, prop4, prop5, prop6, prop7, uniqueOnly, subscriber)
+
+    private fun <A, B, C, D, E, F, G> selectSubscribeInternal(
+            owner: LifecycleOwner?,
+            prop1: KProperty1<S, A>,
+            prop2: KProperty1<S, B>,
+            prop3: KProperty1<S, C>,
+            prop4: KProperty1<S, D>,
+            prop5: KProperty1<S, E>,
+            prop6: KProperty1<S, F>,
+            prop7: KProperty1<S, G>,
+            uniqueOnly: Boolean,
+            subscriber: (A, B, C, D, E, F, G) -> Unit
+    ) = stateStore.observable
+            .map { MvRxTuple7(prop1.get(it), prop2.get(it), prop3.get(it), prop4.get(it), prop5.get(it), prop6.get(it), prop7.get(it)) }
+            .distinctUntilChanged()
+            .subscribeLifecycle(owner, uniqueOnly) { (a, b, c, d, e, f, g) -> subscriber(a, b, c, d, e, f, g) }
+
     private fun <T> Observable<T>.subscribeLifecycle(
         lifecycleOwner: LifecycleOwner? = null,
         uniqueOnly: Boolean,