+++ mvrx/src/main/kotlin/com/airbnb/mvrx/StateContainer.kt
@@ -25,25 +25,6 @@ fun <A : BaseMvRxViewModel<B>, B : MvRxState, C : BaseMvRxViewModel<D>, D : MvRx
 ) = block(viewModel1.state, viewModel2.state, viewModel3.state)
 
 /**
- * Accesses ViewModel state from five ViewModels synchronously and returns the result of the block.
- */
-fun <
-    A : BaseMvRxViewModel<B>, B : MvRxState,
-    C : BaseMvRxViewModel<D>, D : MvRxState,
-    E : BaseMvRxViewModel<F>, F : MvRxState,
-    G : BaseMvRxViewModel<H>, H : MvRxState,
-    I : BaseMvRxViewModel<J>, J : MvRxState,
-    K
-> withState(
-    viewModel1: A,
-    viewModel2: C,
-    viewModel3: E,
-    viewModel4: G,
-    viewModel5: I,
-    block: (B, D, F, H, J) -> K
-) = block(viewModel1.state, viewModel2.state, viewModel3.state, viewModel4.state, viewModel5.state)
-
-/**
  * Accesses ViewModel state from four ViewModels synchronously and returns the result of the block.
  */
 fun <
@@ -59,3 +40,22 @@ fun <
     viewModel4: G,
     block: (B, D, F, H) -> I
 ) = block(viewModel1.state, viewModel2.state, viewModel3.state, viewModel4.state)
+
+/**
+ * Accesses ViewModel state from five ViewModels synchronously and returns the result of the block.
+ */
+fun <
+        A : BaseMvRxViewModel<B>, B : MvRxState,
+        C : BaseMvRxViewModel<D>, D : MvRxState,
+        E : BaseMvRxViewModel<F>, F : MvRxState,
+        G : BaseMvRxViewModel<H>, H : MvRxState,
+        I : BaseMvRxViewModel<J>, J : MvRxState,
+        K
+        > withState(
+        viewModel1: A,
+        viewModel2: C,
+        viewModel3: E,
+        viewModel4: G,
+        viewModel5: I,
+        block: (B, D, F, H, J) -> K
+) = block(viewModel1.state, viewModel2.state, viewModel3.state, viewModel4.state, viewModel5.state)